# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is api for social media application",
        "title": "Social Media API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "paths": {
...
        "/user/find/{userID}": {
            "get": {
                "description": "authentication and find user by id with follow relations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid for user id",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SwaggerResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.SecureUserWithFollowRelations"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/user/finds": {
            "get": {
                "description": "authentication and find users follower relation cursor pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "cursor uuid for user id",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit for users follower relation cursor pagination",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SwaggerResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.UserWithFollowerRelationCursorPagination"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/user/finds-with-fullname": {
            "get": {
                "description": "authentication and find users with fullname cursor pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "search user with fullname",
                        "name": "fullname",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "cursor uuid for user id",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit for users cursor pagination",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SwaggerResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.UserCursorPagination"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/user/toggle-follow/{followingID}": {
            "post": {
                "description": "authentication follow or unfollow and emit follow or unfollow, notification to client",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid for user id following",
                        "name": "followingID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SwaggerResponseWithData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.Follow"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerBadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.SwaggerResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.SecureUser": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "profileBackgroundUrl": {
                    "type": "string"
                },
                "profileUrl": {
                    "type": "string"
                },
                "providerType": {
                    "$ref": "#/definitions/models.ProviderType"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.ProviderType": {
            "type": "string",
            "enum": [
                "LOCAL",
                "GOOGLE",
                "FACEBOOK",
                "GITHUB"
            ],
            "x-enum-varnames": [
                "ProviderTypeLocal",
                "ProviderTypeGoogle",
                "ProviderTypeFacebook",
                "ProviderTypeGithub"
            ]
        },
        "models.Role": {
            "type": "string",
            "enum": [
                "USER",
                "ADMIN"
            ],
            "x-enum-varnames": [
                "RoleUser",
                "RoleAdmin"
            ]
        },
...
        "response.SwaggerBadRequestResponse": {
            "type": "object",
            "properties": {
                "errorFields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.SwaggerResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.SwaggerResponseWithData": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "user.Follow": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "follower": {
                    "$ref": "#/definitions/user.SecureUserFollow"
                },
                "followerId": {
                    "type": "string"
                },
                "following": {
                    "$ref": "#/definitions/user.SecureUserFollow"
                },
                "followingId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "user.Follower": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "follower": {
                    "$ref": "#/definitions/user.SecureUserFollow"
                },
                "followerId": {
                    "type": "string"
                },
                "followingId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "user.FollowerData": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "followerId": {
                    "type": "string"
                },
                "followingId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "user.Following": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "followerId": {
                    "type": "string"
                },
                "following": {
                    "$ref": "#/definitions/user.SecureUserFollow"
                },
                "followingId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "user.SecureUserFollow": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "followers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.FollowerData"
                    }
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "profileBackgroundUrl": {
                    "type": "string"
                },
                "profileUrl": {
                    "type": "string"
                },
                "providerType": {
                    "$ref": "#/definitions/models.ProviderType"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.SecureUserWithFollowRelations": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "followers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.Follower"
                    }
                },
                "followings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.Following"
                    }
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "profileBackgroundUrl": {
                    "type": "string"
                },
                "profileUrl": {
                    "type": "string"
                },
                "providerType": {
                    "$ref": "#/definitions/models.ProviderType"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.SecureUserWithFollowerRelation": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "followers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.Follower"
                    }
                },
                "fullname": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "profileBackgroundUrl": {
                    "type": "string"
                },
                "profileUrl": {
                    "type": "string"
                },
                "providerType": {
                    "$ref": "#/definitions/models.ProviderType"
                },
                "role": {
                    "$ref": "#/definitions/models.Role"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.UserCursorPagination": {
            "type": "object",
            "properties": {
                "nextCursor": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SecureUser"
                    }
                }
            }
        },
        "user.UserWithFollowerRelationCursorPagination": {
            "type": "object",
            "properties": {
                "nextCursor": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.SecureUserWithFollowerRelation"
                    }
                }
            }
        }
    },
...
}